Refactor bp_tpack_loader() to check for BP 1.7.
Previously, we used to check if the current theme supported BuddyPress or
if the theme was bp-default.

We need to also check for BP 1.7 and whether the current theme already
has bp-default templates.

If we're on BP 1.7 and the current theme doesn't have any bp-default
templates, we stop BP TPack from loading as they can now use BP 1.7.

If we're on BP 1.7 and there are bp-default templates, also check if the
theme added add_theme_support( 'buddypress' ).  As BP TPack already adds
this line in bp_tpack_theme_setup(), throw an admin notice.